refactor(infrastructure): migrate core package to infrastructure
